requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
boto3==1.12.34
botocore==1.15.34
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
docutils==0.15.2
flasgger==0.9.4
Flask==1.1.1
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
harvesttext==0.7.3
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jiagu==0.2.3
jieba==0.42.1
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
JPype1==0.7.0
jsonschema==3.2.0
langid==1.1.6
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
networkx==2.4
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyhanlp==0.1.63
pylint==2.4.4
pyparsing==2.4.6
pypinyin==0.37.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-louvain==0.13
pytz==2019.3
PyYAML==5.3.1
rdflib==4.2.2
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
smart-open==1.10.0
tesserocr==2.5.1
textblob==0.15.3
tqdm==4.44.1
typed-ast==1.4.1
urllib3==1.25.8
w3lib==1.21.0
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
